backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
cachetools==5.3.2
anthropic>=0.40.0
gunicorn==21.2.0
httpx>=0.25.0
pymongo==4.7.1
dnspython==2.6.1
